+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Search vulnerability detector alerts for Software SCOs
+
 ## 0.3.0 - 2024-06-09
 
 ### Added